Rename typoed filename
